src/ContentProcessorAPI/pyproject.toml
[project]
name = "contentprocessorapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-appconfiguration>=1.7.1",
    "azure-identity>=1.20.0",
    "azure-storage-blob>=12.24.1",
    "azure-storage-queue>=12.12.0",
    "certifi>=2024.12.14",
    "fastapi[standard]>=0.115.7",
    "poppler-utils>=0.1.0",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.0",
    "pygments>=2.19.1",
    "pymongo>=4.11.1",
    "python-dotenv>=1.0.1",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "ruff>=0.9.3",
]
